fixed the discrepancy between my id types in java and schema. but I have to run docker-compose up twice for the containers to stay up and running like they should. front-end still doesnt work ata ll.

+++ back-end/src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS exercise (
-   id INT NOT NULL AUTO_INCREMENT,
+   id BIGINT NOT NULL AUTO_INCREMENT,
    name varchar(250) NOT NULL,
    exercise_group ENUM('SHOULDERS','CHEST','FOREARMS','OBLIQUES','QUADS','ADDUCTORS','ABS','BICEPS','CARDIO','TRAPS','TRICEPS','ABDUCTORS','HAMSTRINGS','CALVES','LATS','LOWER_BACK','GLUTES') NOT NULL,
    exercise_type ENUM('UPPERBODY', 'LOWERBODY', 'CARDIO', 'ABS') NOT NULL,
@@ -9,20 +9,20 @@ CREATE TABLE IF NOT EXISTS exercise (
 );
 
 CREATE TABLE IF NOT EXISTS equipment (
-   id INT NOT NULL AUTO_INCREMENT,
+   id BIGINT NOT NULL AUTO_INCREMENT,
    name ENUM('DUMBBELLS','BARBELL','FULL_GYM','WEIGHT_BENCH','NONE') NOT NULL,
    PRIMARY KEY (id)
 );
 
 CREATE TABLE IF NOT EXISTS exercise_equipment (
-   exercise_id INT NOT NULL,
-   equipment_id INT NOT NULL,
+   exercise_id BIGINT NOT NULL,
+   equipment_id BIGINT NOT NULL,
    CONSTRAINT fk_exercise FOREIGN KEY (exercise_id) REFERENCES exercise(id) ON UPDATE CASCADE ON DELETE CASCADE,
    CONSTRAINT fk_equipment FOREIGN KEY (equipment_id) REFERENCES equipment(id) ON UPDATE CASCADE ON DELETE CASCADE 
 );
 
 CREATE TABLE IF NOT EXISTS user (
-   id INT NOT NULL AUTO_INCREMENT,
+   id BIGINT NOT NULL AUTO_INCREMENT,
    username VARCHAR(255) NOT NULL UNIQUE,
    password VARCHAR(255) NOT NULL,
    PRIMARY KEY (id)
@@ -30,8 +30,8 @@ CREATE TABLE IF NOT EXISTS user (
 
 CREATE TABLE IF NOT EXISTS workout (
    -- or change the MySQL schema to BIGINT if you prefer Long in Java
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    user_id INT,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT,
     title VARCHAR(255),
     created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
     completed_at DATETIME,
@@ -41,17 +41,17 @@ CREATE TABLE IF NOT EXISTS workout (
 );
 
 CREATE TABLE IF NOT EXISTS workout_exercise (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    workout_id INT,
-    exercise_id INT,
-    exercise_order INT,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    workout_id BIGINT,
+    exercise_id BIGINT,
+    exercise_order BIGINT,
     FOREIGN KEY (workout_id) REFERENCES workout(id),
     FOREIGN KEY (exercise_id) REFERENCES exercise(id)
 );
 
 CREATE TABLE IF NOT EXISTS exercise_set (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    workout_exercise_id INT,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    workout_exercise_id BIGINT,
     set_order INT NOT NULL,
     reps INT,
     weight DECIMAL(5,2) UNSIGNED, -- supports decimal weights like 95.5
